Fix spurious body wake-ups part 1.
Looks like I had added an unnecessary SetAwake in the MT solve.

There's still another problem that can prevent bodies from sleeping when they should. In the MT solve, non-connected bodies can be batched together to avoid the overhead of submitting very small tasks. In order for any bodies in an island to go to sleep, all position constraints in the island must be solved.